Refactor so that report extracting is its own module.
