+++ engine/Extract_Analyze/Summarizer.py
@@ -6,6 +6,8 @@ import pandas as pd
 
 from .Themes import ThemeReader
 from .OutputFolderReader import OutputFolderReader
+from .ReportExtracting import ReportExtractor
+from . import ReferenceChecking
 
 class ReportSummarizer:
     def __init__(self, output_config):
@@ -236,147 +238,4 @@ issues.
         except yaml.YAMLError as exc:
             print(exc)
             return None
-
-        
-
-        
-
-
-class ReportExtractor:
-    def __init__(self, report_text, report_id):
-        self.report_text = report_text
-        self.report_id = report_id
-
-    def extract_important_text(self) -> (str, list):
-        # Get the pages that should be read
-        contents_sections = self.extract_contents_section()
-        if contents_sections == None:
-            print(f'  Could not find contents section in {self.report_id}')
-            return None, None
-        
-        pages_to_read = self.extract_pages_to_read(contents_sections)
-
-        if pages_to_read == None:
-            print(f'  Could not find the findings or analysis section for {self.report_id}')
-            return None, None
-
-        # Retrieve that actual text for the page numbers.
-        print(f"  I am going to be reading these pages: {pages_to_read}")     
-        text = ""
-        for page in pages_to_read: # Loop through the pages and extract the text
-            extracted_text = self.extract_text_between_page_numbers(page, page+1)
-            if extracted_text == None:
-                print(f"  Could not extract text from page {page}")
-                continue
-            text += extracted_text
-
-        return text, pages_to_read
-
-    def extract_text_between_page_numbers(self, page_number_1, page_number_2) -> str:
-        # Create a regular expression pattern to match the page numbers and the text between them
-        pattern = r"<< Page {} >>.*<< Page {} >>".format(page_number_1, page_number_2)
-        matches = re.findall(pattern, self.report_text, re.DOTALL)
-
-
-        if matches:
-            return matches[0]
-        else:
-            # Return everything after the first page number match
-            pattern = r"<< Page {} >>.*".format(page_number_1)
-            matches = re.findall(pattern, self.report_text, re.DOTALL)
-            if matches:
-                return matches[0]
-            else:
-                print("Error: Could not find text between pages " + str(page_number_1) + " and " + str(page_number_2))
-                return None          
-
-    def extract_contents_section(self) -> str:
-        startRegex = r'((Content)|(content)|(Contents)|(contents))([ \w]{0,30}.+)([\n\w\d\sāēīōūĀĒĪŌŪ]*)(.*\.{5,})'
-        endRegex = r'(?<!<< Page \d+ >>[,/.\w\s]*)[\.]{2,} {1,2}[\d]{1,2}'
-
-        # Get the entire string between the start and end regex
-        startMatch = re.search(startRegex, self.report_text)
-        endMatches = list(re.finditer(endRegex, self.report_text))
-        if endMatches:
-            endMatch = endMatches[-1]
-        else:
-            print("Error cant find the end of the contents section")
-            return None
-        
-        if startMatch and endMatch:
-            contents_section = self.report_text[startMatch.start():endMatch.end()]
-        else:
-            return None
-
-        return contents_section
-
-    def extract_pages_to_read(self, content_section) -> list:
-
-        while True: # Repeat until the LLMs gives a valid response
-            try: 
-                # Get 5 responses and only includes pages that are in atleast 3 of the responses
-                model_response = openAICaller.query(
-                        "What page does the analysis start on. What page does the findings finish on? Your response is only a list of integers. No words are allowed in your response. e.g '12,45' or '10,23'. If you cant find the analysis and findings section just return 'None'",
-                        content_section,
-                        temp = 0)
-                
-                if model_response == "None":
-                    return None
-
-                pages_to_read = [int(num) for num in model_response.split(",")]
-
-                # Make the array every page between first and last
-                pages_to_read = list(range(pages_to_read[0], pages_to_read[-1] + 1))
-                break
-            except ValueError:
-                print(f"  Incorrect response from model retrying. \n  Response was: '{model_response}'")
-
-        return pages_to_read
-    
-    def extract_section(self, section_str: str):
-        base_regex_template = lambda section: fr"(( {section}) {{1,3}}(?![\s\S]*^{section}))|((^{section}) {{1,3}})(?![\w\s()]{{1,100}}\.{{2,}})"
-
-        split_section = section_str.split(".")
-        section = split_section[0]
-        endRegex_nextSection = base_regex_template(fr"{int(section)+1}\.1\.?")
-        startRegex = base_regex_template(fr"{int(section)}\.1\.?")
-        endRegexs = [endRegex_nextSection]
-        if len(split_section) > 1:
-            paragraph = split_section[1]
-            endRegex_nextParagraph = base_regex_template(fr"{section}\.{int(paragraph)+1}\.?")
-            endRegexs.insert(0, endRegex_nextParagraph)
-            startRegex = base_regex_template(fr"{section}\.{int(paragraph)}\.?")
-
-        if len(split_section) > 2:
-            sub_paragraph = split_section[2]
-            endRegex_nextSubParagraph = base_regex_template(fr"{section}\.{paragraph}\.{int(sub_paragraph)+1}\.?")
-            endRegexs.insert(0, endRegex_nextSubParagraph)
-            startRegex = base_regex_template(fr"{section}\.{paragraph}\.{int(sub_paragraph)}\.?")
-        
-        # Get the entire string between the start and end regex
-        # Start by looking for just the next subparagraph, then paragraph, then section
-        startMatch = re.search(startRegex, self.report_text, re.MULTILINE)
-
-        endMatch = None
-
-        for endRegex in endRegexs:
-            endMatch = re.search(endRegex, self.report_text, re.MULTILINE)
-            if endMatch:
-                break
-
-        if startMatch == None or endMatch == None:
-            return None
-
-        if endMatch.end() < startMatch.end():
-            print(f"Error: endMatch is before startMatch")
-            print(f"  startMatch: {startMatch.match} \n  endMatch: {endMatch.match}")
-            print(f"  Regexs: {startRegex} \n  {endRegex}")
-            return None
-        
-        if startMatch and endMatch:
-            section_text = self.report_text[startMatch.start():endMatch.end()]
-            return section_text
-
-        print(f"Error: could not find section")
-        return None
